move sockets into components package
