+++ canals/component/__init__.py
@@ -2,3 +2,6 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 from canals.component.component import component, Component
+from canals.component.sockets import InputSocket, OutputSocket
+
+__all__ = ["component", "Component", "InputSocket", "OutputSocket"]
