+++ canals/component/sockets.py
