+++ test/pipeline/unit/test_validation_pipeline_io.py
@@ -5,7 +5,7 @@ import inspect
 
 from canals.pipeline import Pipeline
 from canals.errors import PipelineValidationError
-from canals.pipeline.sockets import InputSocket, OutputSocket
+from canals.component.sockets import InputSocket, OutputSocket
 from canals.pipeline.validation import _find_pipeline_inputs, _find_pipeline_outputs
 from sample_components import Double, AddFixedValue, Sum, Parity
 
@@ -177,6 +177,7 @@ def test_validate_pipeline_input_only_expected_input_is_present():
     with pytest.raises(ValueError, match="The input value of comp2 is already sent by node comp1"):
         pipe.run({"comp1": {"value": 1}, "comp2": {"value": 2}})
 
+
 def test_validate_pipeline_input_only_expected_input_is_present_falsy():
     pipe = Pipeline()
     pipe.add_component("comp1", Double())
@@ -185,16 +186,19 @@ def test_validate_pipeline_input_only_expected_input_is_present_falsy():
     with pytest.raises(ValueError, match="The input value of comp2 is already sent by node comp1"):
         pipe.run({"comp1": {"value": 1}, "comp2": {"value": 0}})
 
+
 def test_validate_pipeline_falsy_input_present():
     pipe = Pipeline()
     pipe.add_component("comp", Double())
     assert pipe.run({"comp": {"value": 0}}) == {"comp": {"value": 0}}
-        
+
+
 def test_validate_pipeline_falsy_input_missing():
     pipe = Pipeline()
     pipe.add_component("comp", Double())
     with pytest.raises(ValueError, match="Missing input: comp.value"):
         pipe.run({"comp": {}})
+
 
 def test_validate_pipeline_input_only_expected_input_is_present_including_unknown_names():
     pipe = Pipeline()
